Refactor: segregate LastPageDiscriminator
